Fix badge-boost properly
There were a few things I got wrong in regards to how I "fixed" the
glitch.

1. I had created a new function that duplicated a lot of the work that the
   current function (that boosts ALL stats in a loop) already had.
   Admittedly, I did this before I really felt comfortable doing the kind
   of refactoring I did in this commit.

2. I wasn't actually taking into account if the player had the badge for
   the stat they were getting boosted. This resulted in getting unearned
   boosts, even if there were only for a single stat.

This commit (and the previous WIP commits associated with it) merge the
logic for boosting a single stat into the function ApplyBadgeStatBoosts.
Now, you can pass that function an argument if you only want a single
stat boosted. I find this solution more elegant than my first. And, of
course, you don't get a boost for a stat you don't have a badge for.